Remove DecisionsEnum and use AdviceType enum instead
