+++ static/decisions/models.py
@@ -2,9 +2,9 @@ import uuid
 
 from django.db import models
 
-from static.decisions.enums import DecisionsEnum
+from cases.enums import AdviceType
 
 
 class Decision(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
-    name = models.CharField(choices=DecisionsEnum.choices, null=False, blank=False, max_length=45)
+    name = models.CharField(choices=AdviceType.choices, null=False, blank=False, max_length=45)
