+++ static/decisions/serializers.py
@@ -1,12 +1,12 @@
 from rest_framework import serializers
 
 from conf.serializers import KeyValueChoiceField
-from static.decisions.enums import DecisionsEnum
+from cases.enums import AdviceType
 from static.decisions.models import Decision
 
 
 class DecisionSerializer(serializers.ModelSerializer):
-    name = KeyValueChoiceField(choices=DecisionsEnum.choices)
+    name = KeyValueChoiceField(choices=AdviceType.choices)
 
     class Meta:
         model = Decision
